feat(indexing): add indexing utilities for hybrid vector and knowledge graph indexes
- Introduced `index_builder.py` for creating hybrid vector indexes and knowledge graph indexes with GPU optimization and async support. This module enhances the indexing capabilities of DocMind AI while adhering to the KISS principle by providing straightforward implementations for both synchronous and asynchronous index creation without unnecessary complexity.
