+++ utils.py
@@ -1,354 +1,23 @@
-"""Utility functions for DocMind AI.
+"""Core utility functions for DocMind AI.
 
-Provides functions for loading documents with LlamaParse extensions for video/audio,
-creating hybrid indexes with ColBERT late-interaction and torch.compile,
-agentic analysis with ReAct agents, and chat with knowledge graphs/multimodal.
+This module contains general-purpose utilities including logging setup,
+hardware detection, and RRF configuration verification.
 
+Functions:
+    setup_logging: Configure application logging.
+    detect_hardware: Detect GPU and FastEmbed providers.
+    verify_rrf_configuration: Verify RRF config against research.
 """
 
-import asyncio
-import base64
-import io
 import logging
-import os
-import tempfile
-from collections.abc import AsyncGenerator
-from typing import Any, ClassVar
+from typing import Any
 
 import torch
-import whisper
-from fastembed import (
-    LateInteractionMultimodalEmbedding,
-    LateInteractionTextEmbedding,
-    SparseTextEmbedding,
-)
-from llama_index.core import (
-    Document,
-    SimpleDirectoryReader,
-    StorageContext,
-    VectorStoreIndex,
-)
-from llama_index.core.agent import ReActAgent
-from llama_index.core.memory import ChatMemoryBuffer
-from llama_index.core.tools import QueryEngineTool, ToolMetadata
-from llama_index.embeddings.fastembed import FastEmbedEmbedding
-from llama_index.llms.ollama import Ollama
-from llama_index.postprocessor.colbert_rerank import ColbertRerank
-from llama_index.vector_stores.qdrant import QdrantVectorStore
-from moviepy.video.io.VideoFileClip import VideoFileClip
-from PIL import Image
-from qdrant_client import AsyncQdrantClient, QdrantClient
-from qdrant_client.models import (
-    Distance,
-    SparseIndexParams,
-    SparseVector,
-    SparseVectorParams,
-    VectorParams,
-)
 
+from model_manager import ModelManager
 from models import AppSettings
 
-settings: AppSettings = AppSettings()
-logger = logging.getLogger(__name__)
-
-
-# Singleton pattern for FastEmbed models to prevent redundant inits
-class FastEmbedModelManager:
-    """Singleton manager for FastEmbed models to prevent redundant initializations."""
-
-    _instance = None
-    _models: ClassVar[dict[str, Any]] = {}
-
-    def __new__(cls) -> "FastEmbedModelManager":
-        """Create singleton instance."""
-        if cls._instance is None:
-            cls._instance = super().__new__(cls)
-        return cls._instance
-
-    def get_text_embedding_model(self, model_name: str | None = None) -> Any:
-        """Get or create a TextEmbedding model.
-
-        Args:
-            model_name: Name of the text embedding model. Defaults to settings value.
-
-        Returns:
-            Cached or new TextEmbedding instance.
-        """
-        model_name = model_name or settings.dense_embedding_model
-        model_key = f"text_{model_name}"
-
-        if model_key not in self._models:
-            from fastembed import TextEmbedding
-
-            self._models[model_key] = TextEmbedding(
-                model_name=model_name,
-                batch_size=settings.embedding_batch_size,
-                providers=["CUDAExecutionProvider", "CPUExecutionProvider"]
-                if settings.gpu_acceleration
-                else ["CPUExecutionProvider"],
-                cache_dir="./embeddings_cache",
-            )
-            logger.info("Created new TextEmbedding model: %s", model_name)
-        else:
-            logger.debug("Reusing cached TextEmbedding model: %s", model_name)
-
-        return self._models[model_key]
-
-    def get_dense_embedding_model(self, model_name: str | None = None) -> Any:
-        """Get or create a dense TextEmbedding model.
-
-        Alias for get_text_embedding_model for better API consistency.
-        """
-        return self.get_text_embedding_model(model_name)
-
-    def get_sparse_embedding_model(self, model_name: str | None = None) -> Any:
-        """Get or create a SparseTextEmbedding model.
-
-        Args:
-            model_name: Name of the sparse embedding model. Defaults to settings value.
-
-        Returns:
-            Cached or new SparseTextEmbedding instance.
-        """
-        model_name = model_name or settings.sparse_embedding_model
-        model_key = f"sparse_{model_name}"
-
-        if model_key not in self._models:
-            self._models[model_key] = SparseTextEmbedding(
-                model_name=model_name,
-                batch_size=settings.embedding_batch_size,
-                providers=["CUDAExecutionProvider", "CPUExecutionProvider"]
-                if settings.gpu_acceleration
-                else ["CPUExecutionProvider"],
-                cache_dir="./embeddings_cache",
-            )
-            logger.info("Created new SparseTextEmbedding model: %s", model_name)
-        else:
-            logger.debug("Reusing cached SparseTextEmbedding model: %s", model_name)
-
-        return self._models[model_key]
-
-    def get_multimodal_embedding_model(
-        self, model_name: str = "Qdrant/colpali-v1.3-fp16"
-    ) -> Any:
-        """Get or create a LateInteractionMultimodalEmbedding model."""
-        model_key = f"multimodal_{model_name}"
-
-        if model_key not in self._models:
-            self._models[model_key] = LateInteractionMultimodalEmbedding(
-                model_name=model_name
-            )
-            logger.info(
-                "Created new LateInteractionMultimodalEmbedding model: %s", model_name
-            )
-        else:
-            logger.debug(
-                "Reusing cached LateInteractionMultimodalEmbedding model: %s",
-                model_name,
-            )
-
-        return self._models[model_key]
-
-    def get_colbert_model(self, model_name: str = "colbert-ir/colbertv2.0") -> Any:
-        """Get or create a LateInteractionTextEmbedding (ColBERT) model.
-
-        Args:
-            model_name: Name of the ColBERT model. Defaults to colbert-ir/colbertv2.0.
-
-        Returns:
-            Cached or new LateInteractionTextEmbedding instance for ColBERT.
-        """
-        model_key = f"colbert_{model_name}"
-
-        if model_key not in self._models:
-            self._models[model_key] = LateInteractionTextEmbedding(
-                model_name=model_name,
-                batch_size=settings.embedding_batch_size,
-                providers=["CUDAExecutionProvider", "CPUExecutionProvider"]
-                if settings.gpu_acceleration
-                else ["CPUExecutionProvider"],
-                cache_dir="./embeddings_cache",
-            )
-            logger.info(
-                "Created new LateInteractionTextEmbedding model: %s", model_name
-            )
-        else:
-            logger.debug(
-                "Reusing cached LateInteractionTextEmbedding model: %s", model_name
-            )
-
-        return self._models[model_key]
-
-    def clear_cache(self) -> None:
-        """Clear all cached models."""
-        self._models.clear()
-        logger.info("FastEmbed model cache cleared")
-
-
-# Global model manager instance
-model_manager = FastEmbedModelManager()
-
-
-def create_tools_from_index(index: dict[str, Any]) -> list[QueryEngineTool]:
-    """Create query engine tools from index.
-
-    Args:
-        index: Dict containing vector and kg indexes.
-
-    Returns:
-        List of QueryEngineTool instances.
-    """
-    # Setup native ColBERT reranker postprocessor with performance monitoring
-    import time
-
-    postprocessors = []
-    try:
-        # Create base ColBERT reranker
-        base_colbert_reranker = ColbertRerank(
-            top_n=settings.reranking_top_k,
-            model="colbert-ir/colbertv2.0",
-            tokenizer="colbert-ir/colbertv2.0",
-            keep_retrieval_score=True,
-        )
-
-        # Wrap with performance monitoring
-        class ColBERTPerformanceMonitor:
-            def __init__(self, base_reranker: Any) -> None:
-                self.base_reranker = base_reranker
-                self.total_queries = 0
-                self.total_reranking_time = 0.0
-
-            def postprocess_nodes(
-                self, nodes: list[Any], query_bundle: Any = None
-            ) -> list[Any]:
-                """Postprocess nodes with performance monitoring."""
-                start_time = time.perf_counter()
-
-                # Track query statistics
-                self.total_queries += 1
-                initial_count = len(nodes) if nodes else 0
-
-                # Perform reranking
-                result = self.base_reranker.postprocess_nodes(nodes, query_bundle)
-
-                # Calculate timing
-                reranking_time = time.perf_counter() - start_time
-                self.total_reranking_time += reranking_time
-
-                # Log performance metrics
-                final_count = len(result) if result else 0
-                avg_time = self.total_reranking_time / self.total_queries
-
-                logging.info(
-                    f"ColBERT reranking: {initial_count}→{final_count} nodes, "
-                    f"time: {reranking_time:.3f}s, avg: {avg_time:.3f}s "
-                    f"(query #{self.total_queries})"
-                )
-
-                return result
-
-        monitored_reranker = ColBERTPerformanceMonitor(base_colbert_reranker)
-        postprocessors.append(monitored_reranker)
-
-        logger.info(
-            "ColBERT reranker enabled: retrieve 20 → rerank to %d "
-            "(Phase 2.2 compliant with performance monitoring)",
-            settings.reranking_top_k,
-        )
-    except Exception as e:
-        logging.warning(f"ColbertRerank postprocessor initialization failed: {e}")
-
-    # Create query engines with optimized parameters (from deep review)
-    vector_query_engine = index["vector"].as_query_engine(
-        similarity_top_k=5,  # Optimized for better relevance (was 20)
-        sparse_top_k=10,  # Dedicated sparse retrieval count
-        hybrid_top_k=8,  # Optimal hybrid fusion count
-        vector_store_query_mode="hybrid",  # Enable hybrid dense+sparse search
-        alpha=settings.rrf_fusion_alpha,  # RRF fusion parameter
-        node_postprocessors=postprocessors,  # Native ColBERT reranking
-    )
-
-    kg_query_engine = index["kg"].as_query_engine(
-        similarity_top_k=10,  # KG queries may need more results
-        include_text=True,  # Include source text with entities
-        node_postprocessors=postprocessors if len(postprocessors) > 0 else None,
-    )
-
-    # Enhanced tools with better descriptions for agent decision-making
-    tools = [
-        QueryEngineTool(
-            query_engine=vector_query_engine,
-            metadata=ToolMetadata(
-                name="hybrid_vector_search",
-                description=(
-                    "Advanced hybrid search combining dense (BGE-Large) and sparse "
-                    "(SPLADE++) embeddings with RRF fusion and ColBERT reranking. "
-                    "Best for: semantic search, document retrieval, finding similar "
-                    "content, answering questions about document content, "
-                    "summarization, and general information extraction. Uses GPU "
-                    "acceleration when available for 100x performance improvement."
-                ),
-            ),
-        ),
-        QueryEngineTool(
-            query_engine=kg_query_engine,
-            metadata=ToolMetadata(
-                name="knowledge_graph_query",
-                description=(
-                    "Knowledge graph search for entity and relationship-based queries. "
-                    "Best for: finding connections between concepts, identifying "
-                    "entities and their relationships, exploring document structure, "
-                    "understanding document hierarchies, and answering questions "
-                    "about how different concepts relate to each other. "
-                    "Complements vector search by providing structured knowledge "
-                    "representation."
-                ),
-            ),
-        ),
-    ]
-
-    return tools
-
-
-def create_agent_with_tools(index: dict[str, Any], llm: Any) -> ReActAgent:
-    """Create a ReActAgent.
-
-    Tools for hybrid search, RRF fusion, and ColBERT reranking.
-
-    Args:
-        index: Dict containing vector and kg indexes with advanced features.
-        llm: The language model instance to use for the agent.
-
-    Returns:
-        ReActAgent configured with enhanced query tools leveraging all advanced features
-    """
-    # Get tools from index
-    tools = create_tools_from_index(index)
-
-    # Create ReActAgent with enhanced memory and error handling
-    try:
-        agent = ReActAgent.from_tools(
-            tools,
-            llm=llm,
-            verbose=True,
-            max_iterations=10,
-            memory=ChatMemoryBuffer.from_defaults(token_limit=8192),
-        )
-
-        logging.info(f"ReActAgent created with {len(tools)} enhanced tools")
-        logging.info("Tools available: hybrid_vector_search, knowledge_graph_query")
-        return agent
-
-    except Exception as e:
-        logging.error(f"ReActAgent creation failed: {e}")
-        # Fallback with basic configuration
-        agent = ReActAgent.from_tools(
-            tools,
-            llm=llm,
-            verbose=True,
-        )
-        logging.warning("Using fallback ReActAgent configuration")
-        return agent
+settings = AppSettings()
 
 
 def setup_logging(log_level: str = "INFO") -> None:
@@ -358,10 +27,6 @@ def setup_logging(log_level: str = "INFO") -> None:
         format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
         handlers=[logging.StreamHandler(), logging.FileHandler("docmind.log")],
     )
-
-
-# Removed custom GPU optimization functions
-# FastEmbed handles GPU acceleration, memory management, and batching natively
 
 
 def detect_hardware() -> dict[str, Any]:
@@ -384,7 +49,7 @@ def detect_hardware() -> dict[str, Any]:
     # Use FastEmbed's native hardware detection
     try:
         # FastEmbed automatically detects available providers
-        test_model = model_manager.get_text_embedding_model("BAAI/bge-small-en-v1.5")
+        test_model = ModelManager.get_text_embedding_model("BAAI/bge-small-en-v1.5")
 
         # Get detected providers from FastEmbed
         try:
@@ -413,540 +78,6 @@ def detect_hardware() -> dict[str, Any]:
         hardware_info["cuda_available"] = torch.cuda.is_available()
 
     return hardware_info
-
-
-async def setup_hybrid_qdrant_async(
-    client: AsyncQdrantClient,
-    collection_name: str,
-    dense_embedding_size: int = 768,
-    recreate: bool = False,
-) -> QdrantVectorStore:
-    """Setup Qdrant with hybrid search support using AsyncQdrantClient.
-
-    Creates or configures a Qdrant collection optimized for hybrid search
-    with both dense and sparse vector support for improved retrieval performance.
-
-    Args:
-        client: AsyncQdrantClient instance for async operations.
-        collection_name: Name of the collection to create/use.
-        dense_embedding_size: Size of dense embeddings (default: 768).
-        recreate: Whether to recreate collection if it exists.
-
-    Returns:
-        QdrantVectorStore configured for hybrid search with RRF fusion.
-    """
-    if recreate and await client.collection_exists(collection_name):
-        await client.delete_collection(collection_name)
-
-    if not await client.collection_exists(collection_name):
-        # Create collection with both dense and sparse vectors
-        # LlamaIndex QdrantVectorStore expects "text-dense" and "text-sparse" names
-        await client.create_collection(
-            collection_name=collection_name,
-            vectors_config={
-                "text-dense": VectorParams(
-                    size=dense_embedding_size,
-                    distance=Distance.COSINE,
-                )
-            },
-            sparse_vectors_config={
-                "text-sparse": SparseVectorParams(
-                    index=SparseIndexParams(
-                        on_disk=False,
-                    )
-                )
-            },
-        )
-
-    # Convert AsyncQdrantClient to sync for QdrantVectorStore compatibility
-    sync_client = QdrantClient(url=settings.qdrant_url)
-
-    return QdrantVectorStore(
-        client=sync_client,
-        collection_name=collection_name,
-        enable_hybrid=True,
-        batch_size=20,
-    )
-
-
-def setup_hybrid_qdrant(
-    client: QdrantClient,
-    collection_name: str,
-    dense_embedding_size: int = 768,
-    recreate: bool = False,
-) -> QdrantVectorStore:
-    """Setup Qdrant with hybrid search support.
-
-    Creates or configures a Qdrant collection optimized for hybrid search
-    with both dense and sparse vector support for improved retrieval performance.
-
-    Args:
-        client: QdrantClient instance for synchronous operations.
-        collection_name: Name of the collection to create/use.
-        dense_embedding_size: Size of dense embeddings (default: 768).
-        recreate: Whether to recreate collection if it exists.
-
-    Returns:
-        QdrantVectorStore configured for hybrid search with RRF fusion.
-    """
-    if recreate and client.collection_exists(collection_name):
-        client.delete_collection(collection_name)
-
-    if not client.collection_exists(collection_name):
-        # Create collection with both dense and sparse vectors
-        # LlamaIndex QdrantVectorStore expects "text-dense" and "text-sparse" names
-        client.create_collection(
-            collection_name=collection_name,
-            vectors_config={
-                "text-dense": VectorParams(
-                    size=dense_embedding_size,
-                    distance=Distance.COSINE,
-                )
-            },
-            sparse_vectors_config={
-                "text-sparse": SparseVectorParams(
-                    index=SparseIndexParams(
-                        on_disk=False,
-                    )
-                )
-            },
-        )
-
-    return QdrantVectorStore(
-        client=client,
-        collection_name=collection_name,
-        enable_hybrid=True,
-        batch_size=20,
-    )
-
-
-def extract_images_from_pdf(pdf_path: str) -> list[dict[str, Any]]:
-    """Extract images from PDF for multimodal processing.
-
-    Args:
-        pdf_path: Path to the PDF file.
-
-    Returns:
-        List of image dictionaries with content and metadata.
-    """
-    images = []
-    try:
-        import fitz  # PyMuPDF
-
-        doc = fitz.open(pdf_path)
-        for page_num in range(len(doc)):
-            page = doc.load_page(page_num)
-            image_list = page.get_images()
-
-            for img_index, img in enumerate(image_list):
-                xref = img[0]
-                pix = fitz.Pixmap(doc, xref)
-
-                if pix.n - pix.alpha < 4:  # GRAY or RGB
-                    img_data = pix.tobytes("ppm")
-                    img_pil = Image.open(io.BytesIO(img_data))
-
-                    # Convert to base64 for storage
-                    buffer = io.BytesIO()
-                    img_pil.save(buffer, format="PNG")
-                    img_base64 = base64.b64encode(buffer.getvalue()).decode()
-
-                    images.append(
-                        {
-                            "image_data": img_base64,
-                            "page_number": page_num + 1,
-                            "image_index": img_index,
-                            "format": "PNG",
-                            "size": img_pil.size,
-                        }
-                    )
-
-                pix = None
-
-        doc.close()
-        logging.info(f"Extracted {len(images)} images from PDF")
-
-    except Exception as e:
-        logging.error(f"PDF image extraction failed: {e}")
-
-    return images
-
-
-def create_native_multimodal_embeddings(
-    text: str,
-    images: list[dict[str, Any]] = None,
-) -> dict[str, Any]:
-    """Use FastEmbed native LateInteractionMultimodalEmbedding.
-
-    Creates multimodal embeddings using FastEmbed's native implementation
-    for optimal performance and memory efficiency.
-
-    Args:
-        text: Text content to embed.
-        images: List of image dictionaries with base64 data. Defaults to None.
-
-    Returns:
-        Dictionary containing native multimodal embeddings with metadata.
-    """
-    embeddings = {
-        "text_embedding": None,
-        "image_embeddings": [],
-        "combined_embedding": None,
-        "provider_used": None,
-    }
-
-    try:
-        # Use FastEmbed native LateInteractionMultimodalEmbedding
-        try:
-            # Use singleton pattern to prevent redundant model initializations
-            model = model_manager.get_multimodal_embedding_model()
-
-            # Process text embedding with native FastEmbed
-            text_embedding = list(model.embed_text([text]))[0]
-            embeddings["text_embedding"] = text_embedding.flatten().tolist()
-
-            # Process image embeddings with native FastEmbed
-            if images:
-                # Save base64 images to temporary files for FastEmbed
-                image_paths = []
-                for i, img_data in enumerate(images):
-                    try:
-                        img_bytes = base64.b64decode(img_data["image_data"])
-                        temp_path = f"{tempfile.gettempdir()}/multimodal_img_{i}.png"
-                        with open(temp_path, "wb") as f:
-                            f.write(img_bytes)
-                        image_paths.append(temp_path)
-                    except Exception as e:
-                        logging.warning(f"Failed to save image {i}: {e}")
-
-                if image_paths:
-                    # Native FastEmbed image embeddings
-                    image_embeddings = list(model.embed_image(image_paths))
-
-                    for i, img_emb in enumerate(image_embeddings):
-                        embeddings["image_embeddings"].append(
-                            {
-                                "embedding": img_emb.flatten().tolist(),
-                                "metadata": images[i] if i < len(images) else {},
-                            }
-                        )
-
-                    # Clean up temporary files
-                    import contextlib
-
-                    for path in image_paths:
-                        with contextlib.suppress(Exception):
-                            os.unlink(path)
-
-            # FastEmbed handles combined embeddings natively
-            embeddings["combined_embedding"] = embeddings["text_embedding"]
-            embeddings["provider_used"] = "fastembed_native_multimodal"
-
-            logging.info("Using FastEmbed native LateInteractionMultimodalEmbedding")
-
-        except ImportError:
-            logging.warning(
-                "FastEmbed LateInteractionMultimodalEmbedding not available, "
-                "using text-only"
-            )
-            # Fallback to FastEmbed text-only
-            fastembed_model = FastEmbedEmbedding(
-                model_name=settings.dense_embedding_model,
-                providers=["CUDAExecutionProvider", "CPUExecutionProvider"],
-                cache_dir="./embeddings_cache",
-            )
-            embeddings["text_embedding"] = fastembed_model.get_text_embedding(text)
-            embeddings["combined_embedding"] = embeddings["text_embedding"]
-            embeddings["provider_used"] = "fastembed_text_only"
-
-    except Exception as e:
-        logging.error(f"Native multimodal embedding creation failed: {e}")
-        # Ultimate fallback to FastEmbed text-only
-        try:
-            fastembed_model = FastEmbedEmbedding(
-                model_name=settings.dense_embedding_model,
-                cache_dir="./embeddings_cache",
-            )
-            embeddings["text_embedding"] = fastembed_model.get_text_embedding(text)
-            embeddings["combined_embedding"] = embeddings["text_embedding"]
-            embeddings["provider_used"] = "fastembed_fallback"
-        except Exception as fallback_e:
-            logging.error(f"All embedding methods failed: {fallback_e}")
-            embeddings["provider_used"] = "failed"
-
-    return embeddings
-
-
-def load_documents_llama(
-    uploaded_files: list[Any], parse_media: bool = False, enable_multimodal: bool = True
-) -> list[Document]:
-    """Load documents using LlamaParse with multimodal support.
-
-    Supports standard document formats plus video/audio ingestion and PDF image
-    extraction for multimodal processing with enhanced error handling.
-
-    Args:
-        uploaded_files: List of uploaded file objects.
-        parse_media: Whether to parse video/audio files. Defaults to False.
-        enable_multimodal: Whether to enable multimodal processing for PDFs.
-            Defaults to True.
-
-    Returns:
-        List of loaded Document objects with multimodal embeddings where applicable.
-
-    Raises:
-        Exception: If critical document loading operations fail.
-    """
-    from llama_parse import LlamaParse
-
-    parser = LlamaParse(result_type="markdown")  # Latest for tables/images
-    docs: list[Document] = []
-
-    for file in uploaded_files:
-        try:
-            with tempfile.NamedTemporaryFile(
-                delete=False, suffix=os.path.splitext(file.name)[1]
-            ) as tmp_file:
-                tmp_file.write(file.getvalue())
-                file_path: str = tmp_file.name
-
-            if parse_media and (
-                file_path.endswith((".mp4", ".avi"))
-                or file_path.endswith((".mp3", ".wav"))
-            ):
-                if "video" in file.type:
-                    clip = VideoFileClip(file_path)
-                    with tempfile.NamedTemporaryFile(
-                        delete=False, suffix=".wav"
-                    ) as audio_tmp:
-                        audio_path = audio_tmp.name
-                    clip.audio.write_audiofile(audio_path)
-                    device = "cuda" if torch.cuda.is_available() else "cpu"
-                    model = whisper.load_model(
-                        "base", device=device
-                    )  # GPU offload if available
-                    result = model.transcribe(audio_path)
-                    text: str = result["text"]
-                    # Extract frames at intervals (e.g., every 5s) per practices
-                    frames: list[Any] = []
-                    for t in range(0, int(clip.duration), 5):
-                        frame = clip.get_frame(t)
-
-                        img = Image.fromarray(frame)
-                        frames.append(img)
-                    docs.append(
-                        Document(
-                            text=text,
-                            metadata={
-                                "images": frames,
-                                "source": file.name,
-                                "type": "video",
-                            },
-                        )
-                    )
-                    os.remove(audio_path)
-                elif "audio" in file.type:
-                    device = "cuda" if torch.cuda.is_available() else "cpu"
-                    model = whisper.load_model("base", device=device)
-                    result = model.transcribe(file_path)
-                    text = result["text"]
-                    docs.append(
-                        Document(
-                            text=text, metadata={"source": file.name, "type": "audio"}
-                        )
-                    )
-            else:
-                # Standard document processing with multimodal support
-                reader = SimpleDirectoryReader(
-                    input_files=[file_path], file_extractor={".*": parser}
-                )
-                loaded_docs = reader.load_data()
-
-                # Enhanced multimodal processing for PDFs
-                if enable_multimodal and file_path.lower().endswith(".pdf"):
-                    try:
-                        # Extract images from PDF
-                        pdf_images = extract_images_from_pdf(file_path)
-
-                        # Process each document with multimodal embeddings
-                        for doc in loaded_docs:
-                            if pdf_images:
-                                # Create local multimodal embeddings (offline)
-                                multimodal_embeddings = (
-                                    create_native_multimodal_embeddings(
-                                        text=doc.text,
-                                        images=pdf_images,
-                                    )
-                                )
-
-                                # Update document with multimodal data
-                                doc.metadata.update(
-                                    {
-                                        "source": file.name,
-                                        "type": "pdf_multimodal",
-                                        "image_count": len(pdf_images),
-                                        "has_images": True,
-                                        "multimodal_embeddings": multimodal_embeddings,
-                                    }
-                                )
-
-                                logging.info(
-                                    f"Created multimodal embeddings for {file.name} "
-                                    f"with {len(pdf_images)} images"
-                                )
-                            else:
-                                # PDF without images - standard text processing
-                                doc.metadata.update(
-                                    {
-                                        "source": file.name,
-                                        "type": "pdf_text_only",
-                                        "has_images": False,
-                                    }
-                                )
-
-                    except Exception as e:
-                        logging.warning(
-                            f"Multimodal processing failed for {file.name}: {e}"
-                        )
-                        # Fallback to standard processing
-                        for doc in loaded_docs:
-                            doc.metadata.update(
-                                {
-                                    "source": file.name,
-                                    "type": "pdf_fallback",
-                                    "has_images": False,
-                                }
-                            )
-                else:
-                    # Standard processing for non-PDF files
-                    for doc in loaded_docs:
-                        doc.metadata.update(
-                            {
-                                "source": file.name,
-                                "type": "standard_document",
-                                "has_images": False,
-                            }
-                        )
-
-                docs.extend(loaded_docs)
-
-            os.remove(file_path)
-
-        except FileNotFoundError as e:
-            logging.error(f"File not found: {file.name} - {str(e)}")
-        except ValueError as e:
-            logging.error(f"Invalid file format: {file.name} - {str(e)}")
-        except Exception as e:
-            logging.error(f"Unexpected error loading {file.name}: {str(e)}")
-
-    logging.info(
-        f"Loaded {len(docs)} documents, multimodal processing enabled: "
-        f"{enable_multimodal}"
-    )
-    return docs
-
-
-async def create_qdrant_hybrid_query_async(
-    client: AsyncQdrantClient,
-    collection_name: str,
-    dense_query: list[float],
-    sparse_query: list[tuple[int, float]],
-    limit: int = 10,
-    rrf_alpha: int = 60,
-) -> list[Any]:
-    """Use Qdrant's native RRF fusion with optimized prefetch mechanism (async version).
-
-    This implements Reciprocal Rank Fusion (RRF) as specified in Phase 2.1:
-    - Uses Qdrant's native RRF fusion for combining dense/sparse results
-    - Implements prefetch mechanism for performance (retrieves limit * 2)
-    - Uses research-backed RRF alpha parameter for rank fusion
-    - Async version for 50-80% performance improvement
-
-    Args:
-        client: AsyncQdrantClient instance.
-        collection_name: Name of the collection.
-        dense_query: Dense query vector.
-        sparse_query: Sparse query vector as (indices, values) tuples.
-        limit: Number of results to return.
-        rrf_alpha: RRF alpha parameter (default 60, from research).
-
-    Returns:
-        Search results with native RRF fusion.
-    """
-    # Convert sparse query to Qdrant format
-    indices, values = zip(*sparse_query, strict=False) if sparse_query else ([], [])
-    sparse_vector = SparseVector(indices=list(indices), values=list(values))
-
-    # Optimized prefetch limits for performance
-    prefetch_limit = max(limit * 2, 20)  # Ensure minimum prefetch for quality
-
-    # Native RRF query with prefetch pattern for performance
-    results = await client.query_points(
-        collection_name=collection_name,
-        prefetch=[
-            {"query": dense_query, "using": "dense", "limit": prefetch_limit},
-            {"query": sparse_vector, "using": "sparse", "limit": prefetch_limit},
-        ],
-        query={"fusion": "rrf"},  # Native RRF fusion!
-        limit=limit,
-    )
-
-    logging.info(
-        f"Async Qdrant native RRF fusion - prefetch: {prefetch_limit}, "
-        f"final: {limit}, alpha: {rrf_alpha}"
-    )
-    return results.points
-
-
-def create_qdrant_hybrid_query(
-    client: QdrantClient,
-    collection_name: str,
-    dense_query: list[float],
-    sparse_query: list[tuple[int, float]],
-    limit: int = 10,
-    rrf_alpha: int = 60,
-) -> list[Any]:
-    """Use Qdrant's native RRF fusion with optimized prefetch mechanism.
-
-    This implements Reciprocal Rank Fusion (RRF) as specified in Phase 2.1:
-    - Uses Qdrant's native RRF fusion for combining dense/sparse results
-    - Implements prefetch mechanism for performance (retrieves limit * 2)
-    - Uses research-backed RRF alpha parameter for rank fusion
-
-    Args:
-        client: QdrantClient instance.
-        collection_name: Name of the collection.
-        dense_query: Dense query vector.
-        sparse_query: Sparse query vector as (indices, values) tuples.
-        limit: Number of results to return.
-        rrf_alpha: RRF alpha parameter (default 60, from research).
-
-    Returns:
-        Search results with native RRF fusion.
-    """
-    # Convert sparse query to Qdrant format
-    indices, values = zip(*sparse_query, strict=False) if sparse_query else ([], [])
-    sparse_vector = SparseVector(indices=list(indices), values=list(values))
-
-    # Optimized prefetch limits for performance
-    prefetch_limit = max(limit * 2, 20)  # Ensure minimum prefetch for quality
-
-    # Native RRF query with prefetch pattern for performance
-    results = client.query_points(
-        collection_name=collection_name,
-        prefetch=[
-            {"query": dense_query, "using": "dense", "limit": prefetch_limit},
-            {"query": sparse_vector, "using": "sparse", "limit": prefetch_limit},
-        ],
-        query={"fusion": "rrf"},  # Native RRF fusion!
-        limit=limit,
-    )
-
-    logging.info(
-        f"Qdrant native RRF fusion - prefetch: {prefetch_limit}, "
-        f"final: {limit}, alpha: {rrf_alpha}"
-    )
-    return results.points
 
 
 def verify_rrf_configuration(settings: AppSettings) -> dict[str, Any]:
@@ -1004,286 +135,3 @@ def verify_rrf_configuration(settings: AppSettings) -> dict[str, Any]:
 
     logging.info(f"RRF Configuration Verification: {verification}")
     return verification
-
-
-async def create_index_async(docs: list[Document], use_gpu: bool) -> dict[str, Any]:
-    """Create hybrid index with Qdrant using async operations.
-
-    Provides 50-80% performance improvement over synchronous operations.
-    ColBERT reranking is handled via native postprocessor in query tools.
-
-    Args:
-        docs: List of documents to index.
-        use_gpu: Whether to use GPU for embeddings.
-
-    Returns:
-        Dict with vector and kg indexes.
-    """
-    try:
-        # Verify RRF configuration meets Phase 2.1 requirements
-        rrf_verification = verify_rrf_configuration(settings)
-        if rrf_verification["issues"]:
-            logging.warning(f"RRF Configuration Issues: {rrf_verification['issues']}")
-            for rec in rrf_verification["recommendations"]:
-                logging.info(f"Recommendation: {rec}")
-
-        # Use AsyncQdrantClient for improved performance
-        async_client = AsyncQdrantClient(url=settings.qdrant_url)
-
-        # Use FastEmbed native GPU acceleration for both dense and sparse embeddings
-        # Dense embedding model with optimized configuration
-        embed_model = FastEmbedEmbedding(
-            model_name=settings.dense_embedding_model,  # BAAI/bge-large-en-v1.5
-            cache_dir="./embeddings_cache",
-            providers=["CUDAExecutionProvider", "CPUExecutionProvider"]
-            if use_gpu and torch.cuda.is_available()
-            else ["CPUExecutionProvider"],
-            batch_size=settings.embedding_batch_size,
-        )
-
-        # Sparse embedding model with optimized configuration
-        from fastembed import SparseTextEmbedding
-
-        sparse_embed_model = SparseTextEmbedding(
-            model_name=settings.sparse_embedding_model,
-            cache_dir="./embeddings_cache",
-            providers=["CUDAExecutionProvider", "CPUExecutionProvider"]
-            if use_gpu and torch.cuda.is_available()
-            else ["CPUExecutionProvider"],
-            batch_size=settings.embedding_batch_size,
-        )
-
-        if use_gpu and torch.cuda.is_available():
-            logging.info(
-                "Using FastEmbed native GPU acceleration for dense and sparse "
-                "embeddings with AsyncQdrantClient"
-            )
-        else:
-            logging.info(
-                "Using FastEmbed CPU mode for embeddings with AsyncQdrantClient"
-            )
-
-        # Setup Qdrant with proper hybrid search configuration using async
-        vector_store = await setup_hybrid_qdrant_async(
-            client=async_client,
-            collection_name="docmind",
-            dense_embedding_size=settings.dense_embedding_dimension,  # BGE-Large 1024D
-            recreate=False,
-        )
-
-        # Create index with RRF-enabled hybrid search
-        index = VectorStoreIndex.from_documents(
-            docs,
-            storage_context=StorageContext.from_defaults(vector_store=vector_store),
-            embed_model=embed_model,
-            sparse_embed_model=sparse_embed_model,
-        )
-
-        # Calculate hybrid_alpha from research-backed weights (dense: 0.7, sparse: 0.3)
-        hybrid_alpha = settings.rrf_fusion_weight_dense / (
-            settings.rrf_fusion_weight_dense + settings.rrf_fusion_weight_sparse
-        )
-
-        # RRF fusion is handled automatically by the hybrid query mode
-        logging.info(
-            f"Using research-backed RRF fusion with weights - "
-            f"Dense: {settings.rrf_fusion_weight_dense:.1f}, "
-            f"Sparse: {settings.rrf_fusion_weight_sparse:.1f}, "
-            f"Hybrid Alpha: {hybrid_alpha:.3f} (AsyncQdrantClient enabled)"
-        )
-
-        # ColBERT reranking handled by native LlamaIndex postprocessor
-        # Simplified per Phase 2.2 requirements
-
-        # Note: KG index creation temporarily disabled for testing (requires OpenAI API)
-        # kg_index = KnowledgeGraphIndex.from_documents(
-        #     docs
-        # )  # For entity/relation queries
-
-        await async_client.close()  # Cleanup async client
-        return index  # Return just the vector index for testing
-    except ValueError as e:
-        logging.error(f"Invalid configuration for index: {str(e)}")
-        raise
-    except Exception as e:
-        logging.error(f"Index creation error: {str(e)}")
-        raise
-
-
-def create_index(docs: list[Document], use_gpu: bool) -> dict[str, Any]:
-    """Create hybrid index with Qdrant, knowledge graph, torch.compile for embeddings.
-
-    ColBERT reranking is handled via native postprocessor in query tools.
-
-    Args:
-        docs: List of documents to index.
-        use_gpu: Whether to use GPU for embeddings.
-
-    Returns:
-        Dict with vector and kg indexes.
-    """
-    try:
-        # Verify RRF configuration meets Phase 2.1 requirements
-        rrf_verification = verify_rrf_configuration(settings)
-        if rrf_verification["issues"]:
-            logging.warning(f"RRF Configuration Issues: {rrf_verification['issues']}")
-            for rec in rrf_verification["recommendations"]:
-                logging.info(f"Recommendation: {rec}")
-
-        client = QdrantClient(url=settings.qdrant_url)
-
-        logging.info(
-            "Using synchronous QdrantClient (consider using create_index_async "
-            "for 50-80%% performance improvement)"
-        )
-
-        # Use FastEmbed native GPU acceleration for both dense and sparse embeddings
-        # Dense embedding model with optimized configuration
-        embed_model = FastEmbedEmbedding(
-            model_name=settings.dense_embedding_model,  # BAAI/bge-large-en-v1.5
-            cache_dir="./embeddings_cache",
-            providers=["CUDAExecutionProvider", "CPUExecutionProvider"]
-            if use_gpu and torch.cuda.is_available()
-            else ["CPUExecutionProvider"],
-            batch_size=settings.embedding_batch_size,
-        )
-
-        # Sparse embedding model with optimized configuration
-        from fastembed import SparseTextEmbedding
-
-        sparse_embed_model = SparseTextEmbedding(
-            model_name=settings.sparse_embedding_model,
-            cache_dir="./embeddings_cache",
-            providers=["CUDAExecutionProvider", "CPUExecutionProvider"]
-            if use_gpu and torch.cuda.is_available()
-            else ["CPUExecutionProvider"],
-            batch_size=settings.embedding_batch_size,
-        )
-
-        if use_gpu and torch.cuda.is_available():
-            logging.info(
-                "Using FastEmbed native GPU acceleration for dense and sparse "
-                "embeddings"
-            )
-        else:
-            logging.info("Using FastEmbed CPU mode for embeddings")
-
-        # Setup Qdrant with proper hybrid search configuration
-        vector_store = setup_hybrid_qdrant(
-            client=client,
-            collection_name="docmind",
-            dense_embedding_size=settings.dense_embedding_dimension,  # BGE-Large 1024D
-            recreate=False,
-        )
-
-        # Create index with RRF-enabled hybrid search
-        index = VectorStoreIndex.from_documents(
-            docs,
-            storage_context=StorageContext.from_defaults(vector_store=vector_store),
-            embed_model=embed_model,
-            sparse_embed_model=sparse_embed_model,
-        )
-
-        # Enhance query engine with RRF fusion - use research-backed weight ratios
-        # Calculate hybrid_alpha from research-backed weights (dense: 0.7, sparse: 0.3)
-        # LlamaIndex hybrid_alpha: 0.0 = full sparse, 1.0 = full dense
-        hybrid_alpha = settings.rrf_fusion_weight_dense / (
-            settings.rrf_fusion_weight_dense + settings.rrf_fusion_weight_sparse
-        )
-
-        # Query engine configuration handled by create_tools_from_index
-
-        # RRF fusion is handled automatically by the hybrid query mode
-        logging.info(
-            f"Using research-backed RRF fusion with weights - "
-            f"Dense: {settings.rrf_fusion_weight_dense:.1f}, "
-            f"Sparse: {settings.rrf_fusion_weight_sparse:.1f}, "
-            f"Hybrid Alpha: {hybrid_alpha:.3f}"
-        )
-
-        # ColBERT reranking handled by native LlamaIndex postprocessor
-        # Simplified per Phase 2.2 requirements
-
-        # Note: KG index creation temporarily disabled for testing (requires OpenAI API)
-        # kg_index = KnowledgeGraphIndex.from_documents(
-        #     docs
-        # )  # For entity/relation queries
-        return index  # Return just the vector index for testing
-    except ValueError as e:
-        logging.error(f"Invalid configuration for index: {str(e)}")
-        raise
-    except Exception as e:
-        logging.error(f"Index creation error: {str(e)}")
-        raise
-
-
-def analyze_documents_agentic(
-    agent: ReActAgent, index: dict[str, Any], prompt_type: str
-) -> str:
-    """Agentic analysis with multi-step reasoning.
-
-    Args:
-        agent: ReActAgent instance.
-        index: Dict of indexes.
-        prompt_type: Type of prompt to use.
-
-    Returns:
-        Analysis response string.
-    """
-    vector_query_engine = index["vector"].as_query_engine()
-    kg_query_engine = index["kg"].as_query_engine()
-
-    tools = [
-        QueryEngineTool(
-            query_engine=vector_query_engine,
-            metadata=ToolMetadata(
-                name="vector_query",
-                description="Query documents using vector similarity search "
-                "for general content retrieval",
-            ),
-        ),
-        QueryEngineTool(
-            query_engine=kg_query_engine,
-            metadata=ToolMetadata(
-                name="knowledge_graph_query",
-                description="Query documents using knowledge graph "
-                "for entity and relationship-based queries",
-            ),
-        ),
-    ]
-    if not agent:
-        # Agent will be properly initialized in app.py with user-selected LLM
-        # This is a fallback that should not be reached in normal operation
-        agent = ReActAgent.from_tools(
-            tools,
-            llm=Ollama(model=settings.default_model),
-            verbose=True,
-        )
-    response = agent.chat(f"Analyze with prompt: {prompt_type}")
-    return response.response  # Parse to AnalysisOutput
-
-
-async def chat_with_agent(
-    agent: ReActAgent, user_input: str, memory: ChatMemoryBuffer
-) -> AsyncGenerator[str, None]:
-    """Async stream chat with agent, supporting KG and multimodal via Gemma/Nemotron.
-
-    Args:
-        agent: ReActAgent instance.
-        user_input: User query string.
-        memory: Chat memory buffer.
-
-    Yields:
-        Chunks of response text.
-    """
-    try:
-        response = await asyncio.to_thread(
-            agent.async_stream_chat, user_input, memory=memory
-        )
-        async for chunk in response.async_response_gen():  # Using async_gen
-            yield chunk
-        # Multimodal handling: If Gemma, use native; for Nemotron (text-only),
-        # extract text features
-    except Exception as e:
-        logging.error(f"Chat generation error: {str(e)}")
-        raise
